visual-testing-agent/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

pydantic==2.6.4
pydantic-settings==2.2.1

opencv-python-headless==4.9.0.80
scikit-image==0.22.0
Pillow==10.2.0
numpy==1.26.4
scipy==1.13.0

sqlalchemy==2.0.29
aiosqlite==0.19.0

python-dotenv==1.0.1

pytesseract==0.3.10

torch>=2.2.0
torchvision>=0.17.0

segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

google-generativeai==0.5.4
groq>=0.5.0

reportlab==4.1.0